+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/advice/CustomExceptionHandler.java
@@ -109,6 +109,14 @@ public class CustomExceptionHandler {
         return map;
     }
 
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(ScheduleAlreadyTakenException.class)
+    public Map<String, String> handleScheduleAlreadyTakenException(ScheduleAlreadyTakenException exception) {
+        Map<String, String> map =  new HashMap<>();
+        map.put("Schedule ERROR", exception.getMessage());
+        return map;
+    }
+